Last commit before merge with master.
Moved stateful state from BaseBuffer to WritableBufferImpl.

Removed overlap check from Memory copyTo() and from UnsafeUtil.

Updated copyTo and compareTo to use isSameResource().
